Unwire the state code and gather it in one place. Refactor as much as is possible.
